requirements.txt

plexapi>=4.15.4
playwright>=1.40.0
python-dotenv>=1.0.0